requirements.txt
paddleocr>=2.7.0
Pillow>=10.0.0
pytesseract>=0.3.10
PyMuPDF>=1.23.0
requests>=2.31.0
pdf2image>=1.16.3
PyQt6>=6.6.0
numpy>=1.24.0
opencv-python>=4.8.0
pytest>=7.4.0
python-dotenv>=1.0.0
pydantic>=2.4.0
httpx>=0.24.0
psycopg2-binary>=2.9.9
SQLAlchemy>=2.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0 